+++ tests/test_optimizer.py
@@ -16,6 +16,7 @@ import desc.examples
 from desc.backend import jit, jnp
 from desc.derivatives import Derivative
 from desc.equilibrium import Equilibrium
+from desc.geometry import FourierRZToroidalSurface
 from desc.grid import LinearGrid
 from desc.objectives import (
     AspectRatio,
@@ -24,6 +25,7 @@ from desc.objectives import (
     FixBoundaryZ,
     FixCurrent,
     FixIota,
+    FixParameter,
     FixPressure,
     FixPsi,
     ForceBalance,
@@ -31,6 +33,7 @@ from desc.objectives import (
     MagneticWell,
     MeanCurvature,
     ObjectiveFunction,
+    PlasmaVesselDistance,
     Volume,
     get_fixed_boundary_constraints,
 )
@@ -315,6 +318,7 @@ def test_no_iterations():
 
 @pytest.mark.unit
 @pytest.mark.slow
+@pytest.mark.optimize
 def test_overstepping():
     """Test that equilibrium is NOT updated when final function value is worse.
 
@@ -400,6 +404,7 @@ def test_overstepping():
 
 @pytest.mark.unit
 @pytest.mark.slow
+@pytest.mark.solve
 def test_maxiter_1_and_0_solve():
     """Test that solves with maxiter 1 and 0 terminate correctly."""
     # correctly meaning they terminate, instead of looping infinitely
@@ -427,6 +432,7 @@ def test_maxiter_1_and_0_solve():
 
 @pytest.mark.unit
 @pytest.mark.slow
+@pytest.mark.solve
 def test_scipy_fail_message():
     """Test that scipy fail message does not cause an error (see PR #434)."""
     eq = Equilibrium(M=3)
@@ -519,14 +525,16 @@ def test_wrappers():
     con_nl = (ForceBalance(eq=eq),)
     obj = ForceBalance(eq=eq)
     with pytest.raises(AssertionError):
-        _ = ProximalProjection(obj, con[0])
+        _ = ProximalProjection(obj, con[0], eq=eq)
     with pytest.raises(AssertionError):
-        _ = ProximalProjection(ObjectiveFunction(con[0]), con[1])
-    with pytest.raises(ValueError):
-        _ = ProximalProjection(ObjectiveFunction(con[0]), ObjectiveFunction(con[1]))
+        _ = ProximalProjection(ObjectiveFunction(con[0]), con[1], eq=eq)
     with pytest.raises(ValueError):
         _ = ProximalProjection(
-            ObjectiveFunction(con[0]), ObjectiveFunction(con + con_nl)
+            ObjectiveFunction(con[0]), ObjectiveFunction(con[1]), eq=eq
+        )
+    with pytest.raises(ValueError):
+        _ = ProximalProjection(
+            ObjectiveFunction(con[0]), ObjectiveFunction(con + con_nl), eq=eq
         )
     ob = ProximalProjection(ObjectiveFunction(con[0]), ObjectiveFunction(con_nl), eq=eq)
     ob.build()
@@ -585,6 +593,7 @@ def test_all_optimizers():
 
 @pytest.mark.slow
 @pytest.mark.regression
+@pytest.mark.optimize
 def test_scipy_constrained_solve():
     """Tests that the scipy constrained optimizer does something.
 
@@ -648,6 +657,7 @@ def test_scipy_constrained_solve():
 
 
 @pytest.mark.unit
+@pytest.mark.solve
 def test_solve_with_x_scale():
     """Make sure we can manually specify x_scale when solving/optimizing."""
     # basically just tests that it runs without error
@@ -857,15 +867,18 @@ def test_auglag():
 
 @pytest.mark.slow
 @pytest.mark.regression
+@pytest.mark.optimize
 def test_constrained_AL_lsq():
     """Tests that the least squares augmented Lagrangian optimizer does something."""
     eq = desc.examples.get("SOLOVEV")
 
     constraints = (
         FixBoundaryR(eq=eq, modes=[0, 0, 0]),  # fix specified major axis position
-        FixPressure(eq=eq),  # fix pressure profile
-        FixIota(eq=eq),  # fix rotational transform profile
-        FixPsi(eq=eq),  # fix total toroidal magnetic flux
+        FixPressure(eq=eq),
+        FixParameter(
+            eq, "i_l", bounds=(eq.i_l * 0.9, eq.i_l * 1.1)
+        ),  # linear inequality
+        FixPsi(eq=eq, bounds=(eq.Psi * 0.99, eq.Psi * 1.01)),  # linear inequality
     )
     # some random constraints to keep the shape from getting wacky
     V = eq.compute("V")["V"]
@@ -900,12 +913,16 @@ def test_constrained_AL_lsq():
     Dwell = constraints[-2].compute_scaled(*constraints[-2].xs(eq2))
     assert (ARbounds[0] - ctol) < AR2 < (ARbounds[1] + ctol)
     assert (Vbounds[0] - ctol) < V2 < (Vbounds[1] + ctol)
+    assert (0.99 * eq.Psi - ctol) <= eq2.Psi <= (1.01 * eq.Psi + ctol)
+    np.testing.assert_array_less((0.9 * eq.i_l - ctol), eq2.i_l)
+    np.testing.assert_array_less(eq2.i_l, (1.1 * eq.i_l + ctol))
     assert eq2.is_nested()
     np.testing.assert_array_less(-Dwell, ctol)
 
 
 @pytest.mark.slow
 @pytest.mark.regression
+@pytest.mark.optimize
 def test_constrained_AL_scalar():
     """Tests that the augmented Lagrangian constrained optimizer does something."""
     eq = desc.examples.get("SOLOVEV")
@@ -946,3 +963,57 @@ def test_constrained_AL_scalar():
     np.testing.assert_allclose(V, V2, atol=ctol, rtol=ctol)
     assert eq2.is_nested()
     np.testing.assert_array_less(-Dwell, ctol)
+
+
+@pytest.mark.slow
+@pytest.mark.unit
+@pytest.mark.optimize
+def test_proximal_with_PlasmaVesselDistance():
+    """Tests that the proximal projection works with fixed surface distance obj."""
+    eq = desc.examples.get("SOLOVEV")
+
+    constraints = (
+        ForceBalance(eq=eq),
+        FixPressure(eq=eq),  # fix pressure profile
+        FixIota(eq=eq),  # fix rotational transform profile
+        FixPsi(eq=eq),  # fix total toroidal magnetic flux
+    )
+    # circular surface
+    a = 2
+    R0 = 4
+    surf = FourierRZToroidalSurface(
+        R_lmn=[R0, a],
+        Z_lmn=[0.0, -a],
+        modes_R=np.array([[0, 0], [1, 0]]),
+        modes_Z=np.array([[0, 0], [-1, 0]]),
+        sym=True,
+        NFP=eq.NFP,
+    )
+
+    grid = LinearGrid(M=eq.M, N=0, NFP=eq.NFP)
+    obj = PlasmaVesselDistance(
+        surface=surf, eq=eq, target=0.5, plasma_grid=grid, surface_fixed=True
+    )
+    objective = ObjectiveFunction((obj,))
+
+    optimizer = Optimizer("proximal-lsq-exact")
+    eq, result = optimizer.optimize(
+        eq,
+        objective,
+        constraints,
+        verbose=3,
+        maxiter=3,
+    )
+
+    # make sure it also works if proximal is given multiple objects in things
+    obj = PlasmaVesselDistance(
+        surface=surf, eq=eq, target=0.5, plasma_grid=grid, surface_fixed=False
+    )
+    objective = ObjectiveFunction((obj,))
+    (eq, surf), result = optimizer.optimize(
+        (eq, surf),
+        objective,
+        constraints,
+        verbose=3,
+        maxiter=3,
+    )